Int64 to Int

+++ test/runtests.jl
@@ -694,7 +694,7 @@ end
 
 # Custom atom type
 mutable struct Person
-    i::Int64
+    i::Int
     status::Status
     mass::Float64
     σ::Float64
